+++ server/storage/objectstorage/log_file_impl_test.go
@@ -41,7 +41,7 @@ func TestNewLogFile(t *testing.T) {
 	assert.Equal(t, "test-segment", logFile.segmentPrefixKey)
 	assert.Equal(t, "test-bucket", logFile.bucket)
 	assert.Equal(t, int64(1000), logFile.buffer.Load().MaxSize)
-	assert.Equal(t, 1024*1024, logFile.maxBufferSize)
+	assert.Equal(t, int64(1024*1024), logFile.maxBufferSize)
 	assert.Equal(t, 1000, logFile.maxIntervalMs)
 }
 
@@ -55,7 +55,7 @@ func TestNewROLogFile(t *testing.T) {
 	assert.Equal(t, "test-segment", logFile.segmentPrefixKey)
 	assert.Equal(t, "test-bucket", logFile.bucket)
 	assert.Nil(t, logFile.buffer.Load())
-	assert.Equal(t, 16*1024*1024, logFile.maxBufferSize)
+	assert.Equal(t, int64(16*1024*1024), logFile.maxBufferSize)
 	assert.Equal(t, 1000, logFile.maxIntervalMs)
 }
